backend/requirements.txt
django
djangorestframework
djongo
pymongo
scikit-learn
pandas
numpy
python-dotenv
djangorestframework-simplejwt
django-cors-headers
PyPDF2
requests
beautifulsoup4
lxml
